fix: add jsdoc
