+++ lib/config/config-loader.js
@@ -13,7 +13,7 @@ const path = require("node:path");
 const fs = require("node:fs/promises");
 const { pathToFileURL } = require("node:url");
 const debug = require("debug")("eslint:config-loader");
-const findFiles = require("./find-files");
+const findFile = require("./find-file");
 const { FlatConfigArray } = require("./flat-config-array");
 const { WarningService } = require("../services/warning-service");
 
@@ -538,7 +538,7 @@ class ConfigLoader {
 			basePath = cwd;
 		} else if (useConfigFile !== false) {
 			debug("Searching for eslint.config.js");
-			configFilePath = findFiles(FLAT_CONFIG_FILENAMES, fromDirectory);
+			configFilePath = findFile(FLAT_CONFIG_FILENAMES, fromDirectory);
 
 			if (configFilePath) {
 				basePath = path.dirname(configFilePath);
